+++ frontend/src/component/common/VerticalTabs/VerticalTab/VerticalTab.tsx
@@ -1,25 +1,40 @@
-import { styled } from '@mui/material';
+import { Button, styled } from '@mui/material';
+import {
+    ITooltipResolverProps,
+    TooltipResolver,
+} from '../../TooltipResolver/TooltipResolver';
 
-const StyledTab = styled('button')<{ selected: boolean }>(
+const StyledTab = styled(Button)<{ selected: boolean }>(
     ({ theme, selected }) => ({
-        cursor: 'pointer',
-        border: 0,
-        backgroundColor: selected
-            ? theme.palette.background.paper
-            : 'transparent',
-        borderLeft: `${theme.spacing(1)} solid ${
-            selected ? theme.palette.primary.main : 'transparent'
-        }`,
-        borderRadius: theme.shape.borderRadiusMedium,
-        padding: theme.spacing(2, 4),
-        color: theme.palette.text.primary,
-        fontSize: theme.fontSizes.bodySize,
-        fontWeight: selected ? theme.fontWeight.bold : theme.fontWeight.medium,
-        textAlign: 'left',
-        transition: 'background-color 0.2s ease',
+        '&.MuiButton-root': {
+            cursor: 'pointer',
+            height: '51px',
+            border: 0,
+            backgroundColor: selected
+                ? theme.palette.background.paper
+                : 'transparent',
+            borderLeft: `${theme.spacing(1)} solid ${
+                selected ? theme.palette.primary.main : 'transparent'
+            }`,
+            borderRadius: theme.shape.borderRadiusMedium,
+            justifyContent: 'start',
+            transition: 'background-color 0.2s ease',
+            color: theme.palette.text.primary,
+            textAlign: 'left',
+            padding: theme.spacing(2, 4),
+            fontSize: theme.fontSizes.bodySize,
+            fontWeight: selected
+                ? theme.fontWeight.bold
+                : theme.fontWeight.medium,
+            lineHeight: 1.2,
+        },
         '&:hover': {
             backgroundColor: theme.palette.neutral.light,
         },
+        '&.Mui-disabled': {
+            pointerEvents: 'auto',
+        },
+        justifyContent: 'space-between',
     })
 );
 
@@ -27,14 +42,26 @@ interface IVerticalTabProps {
     label: string;
     selected?: boolean;
     onClick: () => void;
+    tooltipProps?: Omit<ITooltipResolverProps, 'children'>;
 }
 
 export const VerticalTab = ({
     label,
     selected,
     onClick,
+    tooltipProps,
 }: IVerticalTabProps) => (
-    <StyledTab selected={Boolean(selected)} onClick={onClick}>
-        {label}
-    </StyledTab>
+    <TooltipResolver {...tooltipProps} arrow onClick={e => e.preventDefault()}>
+        <StyledTab
+            selected={Boolean(selected)}
+            onClick={onClick}
+            disableRipple
+            disableElevation
+            disableFocusRipple
+            disableTouchRipple
+            fullWidth
+        >
+            {label}
+        </StyledTab>
+    </TooltipResolver>
 );
