+++ frontend/src/hooks/usePlausibleTracker.ts
@@ -8,7 +8,7 @@ import { EventOptions, PlausibleOptions } from 'plausible-tracker';
  * @see https://plausible.io/docs/custom-event-goals#2-create-a-custom-event-goal-in-your-plausible-analytics-account
  * @example `'download | 'invite' | 'signup'`
  **/
-type CustomEvents = 'invite' | 'upgrade_plan_clicked';
+export type CustomEvents = 'invite' | 'upgrade_plan_clicked';
 
 export const usePlausibleTracker = () => {
     const plausible = useContext(PlausibleContext);
